Format database details
