+++ WooCommerce/Classes/Extensions/Bool+Helpers.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+extension Bool {
+    /// Represents bool value with a string
+    ///
+    var stringRepresentable: String {
+        self ? "✔" : "–"
+    }
+}
