+++ WooCommerce/Classes/Extensions/Int64+Helpers.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+extension Int64 {
+    /// Present the amount for byte count
+    ///
+    var byteCountRepresentable: String {
+        let formatter = ByteCountFormatter()
+        return formatter.string(fromByteCount: self)
+    }
+}
